Book List The BookList class is composed of the book's name, number of pages, author's name, and publication date as its variables.
10 objects are to be created from the BookList class and stored in an ArrayList.

A new Map<String, String> is to be created using the stream structure and lambda expressions, where the book name is mapped to the corresponding author's name.

From this list of 10 Book objects, those with more than 100 pages are to be filtered and returned as a new list. (Stream and lambda expressions can be used.)