Fixture The double round-robin league format will be applied. Each team will play two matches against every other team: one at home and one away.
The left side of the list will indicate the home team, while the right side will indicate the away team.

If an odd number of teams is entered, an additional team named "Bay" will be added to make the number even. A team that is matched with "PASS" will not play a match that week.