Creating Our Own List Class A class is designed to hold data using Java’s generic structure.
The purpose of the class is to maintain a dynamic array, and its data type is meant to be defined dynamically. For this, a generic class should be created.

Class Features:
A default size of 10 should be set for the array inside the class, and its size should be doubled whenever needed.

Two types of constructor methods should be provided for the class:

MyList(): If the empty constructor is used, the initial size of the array should be 10.

MyList(int capacity): The initial size of the array should be set based on the capacity parameter.

The following methods should be implemented:

size(): Returns the number of elements in the array.

getCapacity(): Returns the capacity of the array.

add(T data): Used to add elements to the array of the class. If there is not enough space in the array, the array size should be doubled.

Other Methods:
get(int index): Returns the value at the given index. If an invalid index is provided, null should be returned.

remove(int index): Removes the value at the given index and shifts the remaining elements one position to the left. If the index is invalid, null should be returned.

set(int index, T data): Replaces the element at the given index with a new value. If the index is invalid, null should be returned.

toString(): Returns a string representation of the elements in the list.

Additional Methods:
int indexOf(T data): Returns the index of the specified object in the list. If it is not found, -1 should be returned.

int lastIndexOf(T data): Returns the last index of the specified object in the list. If it is not found, -1 should be returned.

boolean isEmpty(): Indicates whether the list is empty.

T[] toArray(): Returns the elements of the list as an array in the same order.

clear(): Removes all elements from the list, resulting in an empty list.

MyList<T> sublist(int start, int finish): Returns a list for the index range specified in the parameters.

boolean contains(T data): Indicates whether the specified value exists in the array.