OOP KeyStones and Interface Encapsulation
The principle of encapsulation refers to the idea that the variables or attributes of a class can only be accessed and modified through methods that belong to that class. Thanks to this principle, inconsistencies or meaningless operations on objects can be prevented.

It also means that variables are not accessible from outside the class, and how and to what extent variables are defined inside a class is hidden from other parts of the code. By encapsulating our variables, we turn them into a filter that protects them from unwanted situations. Let's try to understand this with an example.

Inheritance

Inheritance in programming works similarly to its real-world concept. When a class inherits from another class, it means that the inheriting class adopts the attributes and behaviors of the other class. If we refer to the class performing the inheritance as the subclass, and the class being inherited from as the superclass, we can say that everything defined in the superclass is also available to the subclass.

Polymorphism

Polymorphism refers to the ability to perform the same task in different ways. An object can carry out the same behavior in various forms and appearances.

Abstraction

In object-oriented programming, the principle of Abstraction means that if it doesn't make sense to instantiate an object from a class, that class can be abstracted. If you want to create a superclass that holds the common features and functionalities of its subclasses but does not have any instance itself, an abstract superclass is created.

Abstraction means defining the core tasks of a class or method without providing the details. Basically, it’s about making a general, simple, and abstract definition of the methods that will be used to solve a problem.

Interfaces

Another way to achieve abstraction in Java is by defining interfaces. Interfaces provide a higher level of abstraction compared to abstract classes because, in an interface, you can only define abstract functions. You cannot define regular functions with method bodies.