Map and Invetory Design A safe house, where health can be restored in the future, and an Inventory store, where currently only weapons can be purchased, have been added to the game.
In later stages, the ability to purchase armor will also be enabled.

Additionally, an Inventory class was created for the acquired weapons, and an aggregation relationship was established with the Player class to manage the player's weapons.

The Inventory class will also be used for more purposes in the future.