IMS (Insurance Management System) There are two types of customer profiles defined for the insurance company: "Individual" and "Enterprise". An abstract class named "Account" is to be designed for the customer profile. The Account class represents the account information where all customer details are kept after logging into the system. Within the Account class, a reference to an object of type "User" is to be included. (This should be modeled as an aggregation relationship.)
The "User" class is used to represent the personal information of the customer. The following fields are to be included in the User class:

First name (String)

Last name (String)

Email (String)

Password (String)

Occupation (String)

Age (int)

Address list (ArrayList<Address>)

Last login date (Date)

In addition, an ArrayList of addresses is to be maintained in the User class. Address information is to be represented in two forms: Home address (HomeAddress) and Business address (BusinessAddress). These address classes are to inherit from an interface called "Address", which should include the functions deemed necessary by the designer.

A class named "AddressManager" is to be created, responsible for adding and removing customer addresses. Two static functions that can add or remove address elements to/from the User object’s address list are to be defined within this class. The names of these functions can be decided by the developer.

In the "Account" class, a final, void method named "showUserInfo" must be defined to print out the user information.

Within the Account class, a list of insurances the customer has purchased is to be maintained. An abstract class named "Insurance" should be defined to represent insurance policies. The class must include:

Insurance name (String)

Insurance cost (double)

Insurance start and end date (Date)

Additionally, an abstract method called "calculate" must be declared. This method is to be implemented in the classes inheriting from it:

HealthInsurance

ResidenceInsurance

TravelInsurance

CarInsurance

Each of these subclasses should override the calculate method and implement their own logic for computing insurance cost.

Given the structure above, the "Account" class must include the following:

A login status of type AuthenticationStatus

A User object

A list of the user's insurances (ArrayList)

An enum called AuthenticationStatus is to be defined, with two constants: SUCCESS and FAIL. SUCCESS should be used when the login is successful, and FAIL when the login fails.

A login method must be defined that accepts an email and password, compares them with the email and password in the User class. If credentials match, login status should be set to SUCCESS. If not, an exception of type InvalidAuthenticationException is to be thrown. This custom exception class should extend Java's Exception class.

Two non-abstract methods should be defined in the Account class:

One for adding an address to the user's address list

One for removing an address from the user's address list

A method returning the current authentication status should also be implemented.

An abstract method must be defined for adding an insurance policy. This method is to be overridden in the Individual and Enterprise subclasses since different profit margins are to be applied when packages are added for individual or corporate customers.

The "Individual" and "Enterprise" classes must inherit from the "Account" class.

A class named "AccountManager" should be implemented. It must maintain a list of Account objects using a TreeSet. Both individual and enterprise accounts should be stored in this set. A login method should be defined in this class, which takes email and password as input, loops through the accounts, and calls the login method on a matching Account object. If a match is found, the corresponding Account object should be returned. Since the login method may throw an InvalidAuthenticationException, a try-catch block must be used.

Since Account objects are to be stored in a TreeSet, the Account class must implement the Comparable interface. The hashCode and equals methods must also be overridden.

Finally, a class that takes email and password input from the keyboard should be created. This class should invoke the login method from the AccountManager class using the provided credentials. If a valid login is detected, the Account object should be returned.