Chat Room A console-based Chat Room Application is to be developed in Java, which is launched from the command line. The application consists of a server and multiple clients that can run on different computers within the same network, such as a Local Area Network (LAN).
A chat room is to be created. Multiple users are to be allowed to connect to the chat room. Messages sent by any user are to be readable by all others.

Upon connecting to the server, a username is to be entered by the user to join the chat. The list of currently online users is to be sent to the newly connected user by the server.

Whenever a new user joins or a user leaves the chat, all users are to be notified accordingly.

The application is to be divided into two parts: server and client, each of which is to be run independently on separate computers.

The following technologies are to be used:

Socket

Thread